deprecated the mplayer and disconnect console commands
